# src/GuildSaber.Website/openapi/GuildSaber.Api.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "GuildSaber.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/login/beatleader": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with BeatLeader",
        "description": "Initiate the BeatLeader authentication flow with optional callback path.",
        "operationId": "BeatLeaderLogin",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login/discord": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with Discord",
        "description": "Initiate the Discord authentication flow with optional callback path.",
        "operationId": "DiscordLogin",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/link/discord": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Initiate linking Discord from authentication flow.",
        "description": "Initiate the Discord authentication flow to link Discord account to an existing authenticated user with optional callback path.",
        "operationId": "DiscordLink",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/auth/callback/link/discord": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Link Discord account to existing authenticated user.",
        "description": "Handles the callback from Discord to link Discord account to existing authenticated user.",
        "operationId": "DiscordLinkCallback",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectHttpResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.2",
                  "title": "Unauthorized",
                  "status": 401,
                  "instance": "GET auth/callback/link/discord",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc4918#section-11.2",
                  "title": "Unprocessable Entity",
                  "status": 422,
                  "detail": null,
                  "instance": "GET auth/callback/link/discord",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/auth/callback/link/discord/redirect": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Link Discord account to existing authenticated user with redirect.",
        "description": "Handles the callback from Discord after authentication to link Discord account to an existing authenticated user and redirects to a specified path from the calling origin with ?type=discordLink&{error}&status as query params.",
        "operationId": "DiscordLinkCallbackWithRedirect",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectHttpResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.2",
                  "title": "Unauthorized",
                  "status": 401,
                  "instance": "GET auth/callback/link/discord/redirect",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc4918#section-11.2",
                  "title": "Unprocessable Entity",
                  "status": 422,
                  "detail": null,
                  "instance": "GET auth/callback/link/discord/redirect",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
    "/auth/callback/beatleader": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Create a session cookie after authenticating with BeatLeader.",
        "description": "Handles the callback from BeatLeader and sets the GuildSaber session cookie.",
        "operationId": "BeatLeaderCallback",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.2",
                  "title": "Unauthorized",
                  "status": 401,
                  "instance": "GET auth/callback/beatleader",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc4918#section-11.2",
                  "title": "Unprocessable Entity",
                  "status": 422,
                  "detail": null,
                  "instance": "GET auth/callback/beatleader",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          },
          "423": {
            "description": "Locked",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "title": "Locked",
                  "status": 423,
                  "detail": null,
                  "instance": "GET auth/callback/beatleader",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": null,
                  "instance": "GET auth/callback/beatleader",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log out the current user by invalidating their session.",
        "operationId": "Logout",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/auth/logout-all": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log out the current user from all sessions by invalidating all their sessions.",
        "operationId": "LogoutAll",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "RedirectHttpResult": {
        "type": "object",
        "properties": {
          "permanent": {
            "type": "boolean"
          },
          "preserveMethod": {
            "type": "boolean"
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          },
          "acceptLocalUrlOnly": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "GuildSaber session cookie. Browsers and cookie-enabled HTTP clients send it automatically.",
        "name": "__Host-GuildSaber.Session",
        "in": "cookie"
      },
      "X-GuildSaber-Request": {
        "type": "apiKey",
        "description": "Required for state-changing requests authenticated with the session cookie.",
        "name": "X-GuildSaber-Request",
        "in": "header"
      }
    }
  },
...
}